pwndbg.qemu.is_remote() needs memoize.reset_on_stop
To see why, consider

  gdb ./vmlinux -ex 'target remote 127.0.0.1:1234'

In this case, it seems qemu.is_remote() is run only once - when
loading "./vmlinux". But at this point, it is not (yet) a remote
debugging session, so "False" is cached. Since no more objfiles
are ever loading in a typical QEMU kernel debugging session,
qemu.is_remote() will be wrong for the rest of the session.